+++ ui/timeline/src/commonMain/kotlin/com/tunjid/heron/timeline/ui/post/PostInteractions.kt
@@ -33,10 +33,8 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.text.BasicText
 import androidx.compose.foundation.text.TextAutoSize
 import androidx.compose.material.icons.Icons
@@ -46,33 +44,19 @@ import androidx.compose.material.icons.rounded.BookmarkBorder
 import androidx.compose.material.icons.rounded.ChatBubbleOutline
 import androidx.compose.material.icons.rounded.Favorite
 import androidx.compose.material.icons.rounded.FavoriteBorder
-import androidx.compose.material.icons.rounded.FormatQuote
 import androidx.compose.material.icons.rounded.Repeat
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.Text
 import androidx.compose.material3.ripple
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.Stable
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.key
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.graphics.vector.rememberVectorPainter
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.text.capitalize
-import androidx.compose.ui.text.intl.Locale
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
@@ -80,7 +64,6 @@ import androidx.compose.ui.unit.sp
 import com.tunjid.composables.ui.animate
 import com.tunjid.heron.data.core.models.Post
 import com.tunjid.heron.data.core.models.Timeline
-import com.tunjid.heron.data.core.models.canQuote
 import com.tunjid.heron.data.core.models.canReply
 import com.tunjid.heron.data.core.models.isBookmarked
 import com.tunjid.heron.data.core.types.PostId
@@ -100,19 +83,10 @@ import com.tunjid.heron.ui.UiTokens.LikeRed
 import com.tunjid.heron.ui.UiTokens.RepostGreen
 import com.tunjid.heron.ui.UiTokens.withDim
 import com.tunjid.heron.ui.rememberLatchedState
-import com.tunjid.heron.ui.sheets.BottomSheetScope
-import com.tunjid.heron.ui.sheets.BottomSheetScope.Companion.ModalBottomSheet
-import com.tunjid.heron.ui.sheets.BottomSheetScope.Companion.rememberBottomSheetState
-import com.tunjid.heron.ui.sheets.BottomSheetState
-import com.tunjid.heron.ui.text.CommonStrings
-import heron.ui.core.generated.resources.cancel
-import heron.ui.core.generated.resources.sign_in
 import heron.ui.timeline.generated.resources.Res
 import heron.ui.timeline.generated.resources.bookmarked
 import heron.ui.timeline.generated.resources.expand_options
 import heron.ui.timeline.generated.resources.liked
-import heron.ui.timeline.generated.resources.quote
-import heron.ui.timeline.generated.resources.remove_repost
 import heron.ui.timeline.generated.resources.reply
 import heron.ui.timeline.generated.resources.repost
 import org.jetbrains.compose.resources.stringResource
@@ -453,201 +427,6 @@ private fun PostInteractionElements(
             animatable.animateTo(PostInteractionButton.BUTTON_MIN_SCALE)
         }
     }
-}
-
-@Stable
-class PostInteractionsSheetState private constructor(
-    isSignedIn: Boolean,
-    scope: BottomSheetScope,
-) : BottomSheetState(scope) {
-    internal var showingAction by mutableStateOf<PostAction.OfInteraction?>(null)
-
-    internal var isSignedIn by mutableStateOf(isSignedIn)
-
-    fun onInteraction(
-        interaction: PostAction.OfInteraction,
-    ) {
-        showingAction = interaction
-    }
-
-    override fun onHidden() {
-        showingAction = null
-    }
-
-    companion object {
-        @Composable
-        fun rememberUpdatedPostInteractionsSheetState(
-            isSignedIn: Boolean,
-            onSignInClicked: () -> Unit,
-            onInteractionConfirmed: (Post.Interaction) -> Unit,
-            // TODO: This should just be an interaction, however the interaction
-            //  is serialized, preventing changing its signature. This should be
-            //  fixed and migrated gracefully when able
-            onQuotePostClicked: (Post.Interaction.Create.Repost) -> Unit,
-        ): PostInteractionsSheetState {
-            val state = rememberBottomSheetState {
-                PostInteractionsSheetState(
-                    isSignedIn = isSignedIn,
-                    scope = it,
-                )
-            }.also { it.isSignedIn = isSignedIn }
-
-            PostInteractionsBottomSheet(
-                state = state,
-                onSignInClicked = onSignInClicked,
-                onInteractionConfirmed = onInteractionConfirmed,
-                onQuotePostClicked = onQuotePostClicked,
-            )
-
-            return state
-        }
-    }
-}
-
-@Composable
-private fun PostInteractionsBottomSheet(
-    state: PostInteractionsSheetState,
-    onSignInClicked: () -> Unit,
-    onInteractionConfirmed: (Post.Interaction) -> Unit,
-    onQuotePostClicked: (Post.Interaction.Create.Repost) -> Unit,
-) {
-    LaunchedEffect(state.showingAction) {
-        when (val interaction = state.showingAction?.interaction) {
-            null -> Unit
-            is Post.Interaction.Create.Repost,
-            is Post.Interaction.Delete.RemoveRepost,
-            -> state.show()
-            is Post.Interaction.Create.Like,
-            is Post.Interaction.Delete.Unlike,
-            is Post.Interaction.Create.Bookmark,
-            is Post.Interaction.Delete.RemoveBookmark,
-            is Post.Interaction.Upsert.Gate,
-            -> {
-                if (state.isSignedIn) {
-                    onInteractionConfirmed(interaction)
-                    state.showingAction = null
-                } else {
-                    state.show()
-                }
-            }
-        }
-    }
-
-    state.ModalBottomSheet {
-        val action = state.showingAction
-        val currentInteraction = action?.interaction
-        Column(
-            modifier = Modifier
-                .padding(horizontal = 16.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            if (state.isSignedIn) {
-                when (currentInteraction) {
-                    is Post.Interaction.Create.Repost,
-                    is Post.Interaction.Delete.RemoveRepost,
-                    -> {
-                        Item(
-                            contentDescription = stringResource(
-                                if (currentInteraction is Post.Interaction.Create.Repost) Res.string.repost
-                                else Res.string.remove_repost,
-                            ),
-                            enabled = true,
-                            icon = Icons.Rounded.Repeat,
-                            onClick = {
-                                onInteractionConfirmed(currentInteraction)
-                                state.hide()
-                            },
-                        )
-                        Item(
-                            contentDescription = stringResource(Res.string.quote),
-                            enabled = action.viewerStats.canQuote,
-                            icon = Icons.Rounded.FormatQuote,
-                            onClick = {
-                                // TODO: Noted in the above, the only valuable part in the
-                                //  interaction is the PostUri weh removing a repost.
-                                //  Stubbing the PostId until a proper migration can be done.
-                                onQuotePostClicked(
-                                    Post.Interaction.Create.Repost(
-                                        postUri = currentInteraction.postUri,
-                                        postId = if (currentInteraction is Post.Interaction.Create.Repost) currentInteraction.postId
-                                        else PostId(""),
-                                    ),
-                                )
-                                state.hide()
-                            },
-                        )
-                    }
-
-                    else -> Unit
-                }
-            }
-
-            // Sheet content
-            OutlinedButton(
-                modifier = Modifier
-                    .fillMaxWidth(),
-                onClick = {
-                    if (!state.isSignedIn) onSignInClicked()
-                    state.hide()
-                },
-                content = {
-                    Text(
-                        text = stringResource(
-                            if (state.isSignedIn) CommonStrings.cancel
-                            else CommonStrings.sign_in,
-                        ),
-                        style = MaterialTheme.typography.bodyLarge,
-                    )
-                },
-            )
-            Spacer(
-                Modifier.height(16.dp),
-            )
-        }
-    }
-}
-
-@Composable
-private fun Item(
-    contentDescription: String,
-    enabled: Boolean,
-    icon: ImageVector,
-    onClick: () -> Unit,
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(CircleShape)
-            .then(
-                when {
-                    enabled -> Modifier.clickable(onClick = onClick)
-                    else -> Modifier.alpha(0.6f)
-                },
-            )
-            .padding(
-                horizontal = 8.dp,
-                vertical = 8.dp,
-            )
-            .semantics {
-                this.contentDescription = contentDescription
-            },
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        content = {
-            Icon(
-                modifier = Modifier
-                    .size(24.dp),
-                imageVector = icon,
-                contentDescription = null,
-            )
-            Text(
-                modifier = Modifier,
-                text = contentDescription
-                    .capitalize(locale = Locale.current),
-                style = MaterialTheme.typography.bodyLarge,
-            )
-        },
-    )
 }
 
 private val Timeline.Presentation.postInteractionArrangement: Arrangement.Horizontal
