+++ ui/timeline/src/commonMain/kotlin/com/tunjid/heron/timeline/utilities/SheetsViewModelInitializers.kt
@@ -2,6 +2,7 @@ package com.tunjid.heron.timeline.utilities
 
 import com.tunjid.heron.timeline.ui.sheets.embedrecordoptions.EmbeddableRecordOptionsViewModelInitializer
 import com.tunjid.heron.timeline.ui.sheets.mutedwords.MutedWordsViewModelInitializer
+import com.tunjid.heron.timeline.ui.sheets.postinteractions.PostInteractionsViewModelInitializer
 import com.tunjid.heron.timeline.ui.sheets.postoptions.PostOptionsViewModelInitializer
 import com.tunjid.heron.timeline.ui.sheets.selectlist.SelectListViewModelInitializer
 import com.tunjid.heron.timeline.ui.sheets.threadgate.ThreadGateViewModelInitializer
@@ -12,4 +13,5 @@ class SheetsViewModelInitializers(
     val threadGateViewModelInitializer: ThreadGateViewModelInitializer,
     val embeddableRecordOptionsViewModelInitializer: EmbeddableRecordOptionsViewModelInitializer,
     val selectListViewModelInitializer: SelectListViewModelInitializer,
+    val postInteractionsViewModelInitializer: PostInteractionsViewModelInitializer,
 )
